requirements.txt
python-gitlab>=3.10.0